requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0rc9
ffmpeg-python==0.2.0
filelock==3.8.0
fire==0.4.0
future==0.18.2
h11==0.14.0
huggingface-hub==0.10.0
more-itertools==8.14.0
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pip==22.2.2
pyparsing==3.0.9
PySocks==1.7.1
python-random-name-generator==0.0.6
PyYAML==6.0
randomname==0.1.5
regex==2022.9.13
requests==2.28.1
selenium==4.5.0
setuptools==65.4.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
termcolor==2.0.1
tokenizer==3.4.2
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
trio-websocket==0.9.2
trio==0.22.0
typing-extensions==4.4.0
urllib3==1.26.12
v==0.0.0
wsproto==1.2.0
